Finalisierung

+++ 10-PDE/partdiff.h
@@ -36,7 +36,7 @@
 
 struct options
 {
-	int32_t term_iteration; /* terminate if iteration number reached          */
+	uint64_t term_iteration; /* terminate if iteration number reached          */
 	uint64_t number;         /* Number of threads                              */
 	uint64_t method;         /* Gauss Seidel or Jacobi method of iteration     */
 	uint64_t interlines;     /* matrix size = interlines*8+9                   */
